Relocation of files
