+++ app/.vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "vs-code-prettier-eslint.prettierLast": true,
+    "eslint.format.enable": true
+}
