+++ app/src/Components/Screens/Company/Company.tsx
@@ -2,8 +2,8 @@ import type { NativeStackScreenProps } from '@react-navigation/native-stack';
 import type { PropsWithChildren } from 'react';
 import React from 'react';
 import { Title } from '../../../Helpers/StylizedComponents';
-import WrapperView from '../../../Helpers/WrapperView';
 import type { CompanyNavigationRoutesType } from '../../../types/NavigationRoutesType';
+import WrapperView from '../../Common/WrapperView';
 
 type CompanyScreenProps = NativeStackScreenProps<
   CompanyNavigationRoutesType,
